refactor: Introduce Dropdown with date formatting
